refactor: Change folder structure and component names
