+++ src/pages/Exercise02/assets/mountains.jpeg
