+++ src/pages/MoviesShop/assets/styles.css
@@ -1,4 +1,4 @@
-.exercise01 {
+.moviesShop {
   background-color: var(--black);
   color: var(--white);
   height: 100vh;
